Changes to programs which didn't compile

+++ ft_strncpy.c
@@ -6,7 +6,7 @@
 /*   By: rysmith <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/12 14:22:21 by rysmith           #+#    #+#             */
-/*   Updated: 2019/06/12 14:32:10 by rysmith          ###   ########.fr       */
+/*   Updated: 2019/07/15 14:40:23 by rysmith          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strncpy(char *dst, const char *src, size_t len)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (src[i] != '\0')
